backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
pydantic[email]==2.5.0
email-validator==2.1.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
numpy==1.25.2
opencv-python-headless==4.8.1.78
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
psutil==5.9.6